refactor(app-shell): add more code split entry points
