+++ .changeset/spotty-points-cross.md
@@ -1,0 +1,6 @@
+---
+'@commercetools-frontend/application-shell': patch
+'@commercetools-local/playground': patch
+---
+
+Add more code split entry points
