+++ packages/application-shell/src/components/authenticated/authenticated.tsx
@@ -1,10 +1,10 @@
 import { Route, Switch } from 'react-router-dom';
 import type { ApplicationWindow } from '@commercetools-frontend/constants';
 import type { TAsyncLocaleDataProps } from '@commercetools-frontend/i18n';
-
+import SuspendedRoute from '../suspended-route';
 import AmILoggedIn from './am-i-logged-in';
 import hasCachedAuthenticationState from './has-cached-authentication-state';
-import OidcCallback from './oidc-callback';
+import OidcCallback from './oidc-callback.async';
 
 declare let window: ApplicationWindow;
 
@@ -38,18 +38,20 @@ Authenticated.displayName = 'Authenticated';
 
 const AuthenticationRoutes = (props: TAuthenticatedProps) => (
   <Switch>
-    <Route path={`/account/oidc/callback`}>
+    <SuspendedRoute path={`/account/oidc/callback`}>
       <OidcCallback
         locale={props.locale}
         applicationMessages={props.applicationMessages}
       />
-    </Route>
-    <Route path={`/:projectKey/${window.app.entryPointUriPath}/oidc/callback`}>
+    </SuspendedRoute>
+    <SuspendedRoute
+      path={`/:projectKey/${window.app.entryPointUriPath}/oidc/callback`}
+    >
       <OidcCallback
         locale={props.locale}
         applicationMessages={props.applicationMessages}
       />
-    </Route>
+    </SuspendedRoute>
     <Route>
       <Authenticated {...props} />
     </Route>
