+++ packages/application-shell/src/components/authenticated/oidc-callback.async.ts
@@ -1,0 +1,7 @@
+import { lazy } from 'react';
+
+const OidcCallback = lazy(
+  () => import('./oidc-callback' /* webpackChunkName: "oidc-callback" */)
+);
+
+export default OidcCallback;
