+++ packages/application-shell/src/components/locale-switcher/index.ts
@@ -1,1 +1,7 @@
-export { default } from './locale-switcher';
+import { lazy } from 'react';
+
+const LocaleSwitcher = lazy(
+  () => import('./locale-switcher' /* webpackChunkName: "locale-switcher" */)
+);
+
+export default LocaleSwitcher;
