+++ packages/application-shell/src/components/project-switcher/index.ts
@@ -1,1 +1,7 @@
-export { default } from './project-switcher';
+import { lazy } from 'react';
+
+const ProjectSwitcher = lazy(
+  () => import('./project-switcher' /* webpackChunkName: "project-switcher" */)
+);
+
+export default ProjectSwitcher;
