+++ packages/application-shell/src/components/redirect-to-login/index.ts
@@ -1,0 +1,8 @@
+import { lazy } from 'react';
+
+const RedirectToLogin = lazy(
+  () =>
+    import('./redirect-to-login' /* webpackChunkName: "redirect-to-login" */)
+);
+
+export default RedirectToLogin;
