+++ packages/application-shell/src/components/redirect-to-login/redirect-to-login.tsx
@@ -8,11 +8,11 @@ import {
 } from '@commercetools-frontend/url-utils';
 import { OIDC_RESPONSE_TYPES } from '../../constants';
 import useIsServedByProxy from '../../hooks/use-is-served-by-proxy';
+import getMcOrigin from '../../utils/get-mc-origin';
 import { buildOidcScope } from '../../utils/oidc';
 import * as oidcStorage from '../../utils/oidc-storage';
 import type { AuthorizeSessionState } from '../authenticated/types';
 import Redirector from '../redirector';
-import { getMcOrigin } from './helpers';
 
 declare let window: ApplicationWindow;
