+++ packages/application-shell/src/components/redirect-to-logout/index.ts
@@ -1,0 +1,8 @@
+import { lazy } from 'react';
+
+const RedirectToLogout = lazy(
+  () =>
+    import('./redirect-to-logout' /* webpackChunkName: "redirect-to-logout" */)
+);
+
+export default RedirectToLogout;
