+++ packages/application-shell/src/utils/get-mc-origin/get-mc-origin.spec.ts
@@ -1,4 +1,4 @@
-import { getMcOrigin } from './helpers';
+import getMcOrigin from './get-mc-origin';
 
 describe.each`
   mcApiUrl                                               | mcUrl                                                   | realMcUrl
