+++ packages/application-shell/src/utils/get-mc-origin/index.ts
@@ -1,0 +1,1 @@
+export { default } from './get-mc-origin';
