+++ playground/src/components/echo-server/index.js
@@ -1,1 +1,7 @@
-export { default } from './echo-server';
+import { lazy } from 'react';
+
+const EchoServer = lazy(() =>
+  import('./echo-server' /* webpackChunkName: "echo-server" */)
+);
+
+export default EchoServer;
