+++ playground/src/components/formatters-demo/index.js
@@ -1,1 +1,7 @@
-export { default } from './formatters-demo';
+import { lazy } from 'react';
+
+const FormattersDemo = lazy(() =>
+  import('./formatters-demo' /* webpackChunkName: "formatters-demo" */)
+);
+
+export default FormattersDemo;
