+++ playground/src/components/notifications-playground/index.js
@@ -1,1 +1,9 @@
-export { default } from './notifications-playground';
+import { lazy } from 'react';
+
+const NotificationsPlayground = lazy(() =>
+  import(
+    './notifications-playground' /* webpackChunkName: "notifications-playground" */
+  )
+);
+
+export default NotificationsPlayground;
