+++ playground/src/components/state-machines-list/index.js
@@ -1,1 +1,7 @@
-export { default } from './state-machines-list';
+import { lazy } from 'react';
+
+const StateMachinesList = lazy(() =>
+  import('./state-machines-list' /* webpackChunkName: "state-machines-list" */)
+);
+
+export default StateMachinesList;
